+++ src/util/clinkedlist.h
@@ -70,6 +70,6 @@ pcb_t* find_process(CircularList* list, pid_t pid);
  * @return pcb_t* Pointer to the found process control block, or NULL if not
  * found.
  */
-pcb_t* find_process_job_id(CircularList* list, u_int64_t index);
+pcb_t* find_process_job_id(CircularList* list, int index);
 
 #endif  // SCHEDULER_LIST_H
