+++ src/util/kernel.h
@@ -113,7 +113,7 @@ typedef struct pcb_t {
   bool is_bg; /** @brief To signal whether or not the processor in background
                                   is terminated */
 
-  u_int64_t job_num; /** @brief The job number if process is in background
+  int job_num; /** @brief The job number if process is in background
                                   or stopped */
 
   char* cmd_name; /** @brief The command name that this processor has
